Added DyeableAirplaneEntity class
EconomyPlaneEntity and ScarletBiplaneEntity now extend DyeableAirplaneEntity, this class is responsible for holding all of the new changes common to both classes
